+++ src/components/MTableActions/index.js
@@ -3,8 +3,10 @@ import PropTypes from 'prop-types';
 
 function MTableActions({
   actions,
+  columns,
   components,
-  data,
+  data = {},
+  onColumnsChanged,
   size,
   disabled,
   forwardedRef
@@ -18,8 +20,10 @@ function MTableActions({
         <components.Action
           action={action}
           key={'action-' + index}
+          columns={columns}
           data={data}
           size={size}
+          onColumnsChanged={onColumnsChanged}
           disabled={disabled}
         />
       ))}
@@ -27,12 +31,8 @@ function MTableActions({
   );
 }
 
-MTableActions.defaultProps = {
-  actions: [],
-  data: {}
-};
-
 MTableActions.propTypes = {
+  columns: PropTypes.array,
   components: PropTypes.object.isRequired,
   actions: PropTypes.array.isRequired,
   data: PropTypes.oneOfType([
@@ -40,6 +40,7 @@ MTableActions.propTypes = {
     PropTypes.arrayOf(PropTypes.object)
   ]),
   disabled: PropTypes.bool,
+  onColumnsChanged: PropTypes.func,
   size: PropTypes.string,
   forwardedRef: PropTypes.element
 };